+++ BojLeave/BojLeave.Api/Controllers/AuthController.cs
@@ -1,4 +1,4 @@
-using BojLeave.Application.Auth;
+using BojLeave.Application;
 using Microsoft.AspNetCore.Mvc;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
