+++ BojLeave/BojLeave.Api/Middleware/ServiceCollectionExtensions.cs
@@ -25,8 +25,8 @@ namespace BojLeave.Api.Middleware
         public static IServiceCollection AddBojLeaveCoreServices(this IServiceCollection services)
         {
             // Register core application services here
-            services.AddScoped<BojLeave.Application.Auth.ILoginService, BojLeave.Application.Auth.LoginService>();
-            services.AddSingleton<BojLeave.Application.Auth.IJwtTokenGenerator, BojLeave.Application.Auth.JwtTokenGenerator>();
+            services.AddScoped<BojLeave.Application.ILoginService, BojLeave.Application.LoginService>();
+            services.AddSingleton<BojLeave.Application.IJwtTokenGenerator, BojLeave.Application.JwtTokenGenerator>();
             // Register SQL Server DbContext
             services.AddDbContext<BojLeave.Infrastructure.BojLeaveDbContext>(options =>
                 options.UseSqlServer("YourConnectionStringHere"));
