+++ BojLeave/BojLeave.Api/Models/LoginRequest.cs
@@ -1,0 +1,8 @@
+namespace BojLeave.Api.Models
+{
+    public class LoginRequest
+    {
+        public string? Username { get; set; }
+        public string? Password { get; set; }
+    }
+}
