+++ BojLeave/BojLeave.Api/Models/RefreshTokenRequest.cs
@@ -1,0 +1,7 @@
+namespace BojLeave.Api.Models
+{
+    public class RefreshTokenRequest
+    {
+        public string? Token { get; set; }
+    }
+}
