+++ BojLeave/BojLeave.Application/Services/IJwtTokenGenerator.cs
@@ -1,0 +1,10 @@
+using System.Security.Claims;
+
+namespace BojLeave.Application
+{
+    public interface IJwtTokenGenerator
+    {
+        string GenerateToken(string username, IEnumerable<Claim> claims);
+        string RefreshToken(string token);
+    }
+}
